+++ examples/Primitives/src/main.cpp
@@ -33,7 +33,7 @@ int main(int argc, char *argv[]) {
     auto messageBus = new JamJar::MessageBus();
     new JamJar::EntityManager(messageBus);
     auto game = new Primitives(messageBus);
-    new JamJar::Standard::_2D::WebGL2System(messageBus, context);
+    new JamJar::Standard::_2D::WebGL2System(messageBus, window, context);
     new JamJar::Standard::_2D::InterpolationSystem(messageBus);
     new JamJar::Standard::_2D::PrimitiveSystem(messageBus);
     new JamJar::Standard::SDL2InputSystem(messageBus);